* created a separate file for the ActionArea view * processing created_at date string value to display the relative date in the Action area
