+++ NewsAppDemo/NewsAppDemo/Models/PostModel.swift
@@ -22,4 +22,5 @@ struct Post: Decodable, Identifiable {
     let points: Int
     let title: String
     let url: String?
+    let created_at: String
 }
