+++ NewsAppDemo/NewsAppDemo/Views/NewsCell.swift
@@ -13,9 +13,12 @@ struct NewsCell: View {
     let urlString: String?
     let id: String
     let title: String
+    let dateCreated: String
     
     var body: some View {
-        VStack {
+        print("date created = ", self.dateCreated)
+        
+        return VStack {
             ZStack(alignment: .leading) {
                 
                 Image("testImage")
@@ -43,52 +46,14 @@ struct NewsCell: View {
                 }.buttonStyle(PlainButtonStyle())
             }
             
-            ActionArea()
-        }
-    }
-}
-
-struct ActionAreaButton: View {
-    
-    let iconName: String
-    let action: () -> ()
-    
-    var body: some View {
-        Button(action: action) {
-            Image(systemName: iconName)
-                .font(.callout)
-                .foregroundColor(.black)
-        }
-        
-        .padding(10)
-    }
-}
-
-struct ActionArea: View {
-    
-    var body: some View {
-        HStack {
-            Text("2h")
-                .font(.callout)
-                .foregroundColor(.red)
-                .padding(.leading)
-            Spacer()
-            ActionAreaButton(iconName: "square.and.arrow.up", action: {})
-            ActionAreaButton(iconName: "bookmark", action: {}).padding(.trailing)
-        }
-        .frame(height: 50)
-        .background(Color.clear)
-        .padding(EdgeInsets(top: -5, leading: 0, bottom: -5, trailing: 0))
-        
-        .onTapGesture {
-            // disables tap gesture on Action Area
+            ActionArea(dateCreatedString: dateCreated)
         }
     }
 }
 
 struct NewsCell_Previews: PreviewProvider {
     static var previews: some View {
-        NewsCell(urlString: "urlString", id: "id", title: "slkfjsd slkjdf slkjsdf sdlfjs fsfsjfjls")
+        NewsCell(urlString: "urlString", id: "id", title: "test title", dateCreated: "date" )
             .previewLayout(.sizeThatFits)
     }
 }
