+++ NewsAppDemo/NewsAppDemo/Views/NewsListView.swift
@@ -28,7 +28,7 @@ struct NewsListView: View {
     var body: some View {
         NavigationView {
             List(newsListViewModel.newsList){ post in
-                NewsCell(urlString: post.url, id: post.id, title: post.title).buttonStyle(PlainButtonStyle())
+                NewsCell(urlString: post.url, id: post.id, title: post.title, dateCreated: post.created_at ).buttonStyle(PlainButtonStyle())
             }.padding(EdgeInsets(top: .zero, leading: offsetEdgePadding, bottom: .zero, trailing: offsetEdgePadding))
             
             .navigationBarTitle(
